Revert seed change in TestUtilities.RNG
Although a random seed is good for testing a wide range of possible
inputs, it creates the problem of not knowing the inputs unless the
random seed is logged. Instead of logging, a constant seed is set and if
needed, more tests are run, further expanding use of the one seed.

To support the use of a singular seed, RandomString has been rewritten.