Fix off-by-one bug in ItemTest
Before, the test would find an Item in the database through a chosen ID
generated by RNG. However, this caused an off-by-one error and could
also cause future problems if the ITEM_LIST is randomly arranged.

Instead, index into the ITEM_LIST and get the ID from an Item instead.