Change names of tests in ItemTests
Test names are now more descriptive as to what unit of work is being
tested, the conditions of testing, and the expected result.

In addition, for the edit and delete tests, the modification operation
and assertion after have been divided into two different "using"
contexts, to more accurately represent two seperate operations.