Create tests for Item page methods
For any POST method, the object in question is manually defined, instead
of relying on supplying form data within the HTTP request and using
BindProperty to define the object.

In addition, tests for POST methods must check the database context to
see if the item was truly unsaved. In order to do this, a custom query
must be made comparing values other than the ID, as ID for the new Item
should have never been set.

Tests still need to be written for EditModel and DeleteModel.