Create unit tests for ItemReview
Seeding data fields in BergTestDataTemplate have been changed to
instance from non-static. It should be this way because the class is
recreated per test. Will likely need to create static data in another
abstract class.

GetAll test still fails, as the navigation fields of the foreign keys
are null. However, the Id fields for the foreign keys are present, so
this may not pose a problem. However, it would be good to understand why
Entity Framework is not retrieving and inserting the correct objects.