Change ItemReview foreign key dependency to Ids
ItemReview objects should now be created with foreign keys defined using
an Id instead of the actual foreign key object. This way, an API call to
create an ItemReview does not require a search query to populate the
foreign key objects, only Ids which need to be verified.

Added definitions for Equals, GetHashcode, and ToString for ItemReview.

Created test case for checking lazy loading of foreign key objects for
ItemReview.